Move tests for different classes into separate files
